+++ src/main/java/com/phamxuantoan/webshop/repository/PermissionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermissionRepository extends JpaRepository<PermissionEntity, Integer> {
+public interface PermissionRepository extends JpaRepository<PermissionEntity, Long> {
 }
